docker/requirements.txt
numpy==1.24.4
torch==2.1.2
transformers==4.36.2
Pillow==9.5.0
scikit-learn==1.2.2
joblib==1.2.0
streamlit